feat: implement core business logic for loans, reservations and inventory
- Implement LoanService with strict business rules (limit per user, overdue checks, copy availability)
- Add returnBook functionality updating physical copy condition and return date
- Implement scheduled task (@Scheduled) to automatically mark overdue loans
- Implement ReservationService with validation for historical vs active reservations
- Implement BookCopyService with batch insert optimization for inventory
- Enable scheduling in LibraryApiApplication

+++ src/main/java/com/app/models/dtos/BookDTO.java
@@ -17,6 +17,6 @@ public class BookDTO {
     private String description;
     private LocalDate publicationDate;
     private String genre;
-    private int bookCopies;
+    private Long bookCopies;
     private List<AuthorDTO> authors;
 }
